Initial commit of the vagrant testing infrastructure
THe plan with this is to make it such that you can run one command to
run unit tests on all supported platforms. This is going to be
accomplished by developing on OS X and running tests on various flavors
of Linux Vagrant virtual machines. They obviously need to be properly
configured to compile osquery, so that where `devtools/cookbooks` comes
in.