initialize schedule to a consistent time
With this change, queries that are scheduled to run with a certain
interval will be consistently ran at a consistent time of the day,
modified only by their splay, regardless of when you started / reloaded
the osqueryd process.

Explicitly, queries that are to be ran once a day will run, splayed, at
midnight. Queries set to run every four hours will run at 0:00, 4:00,
8:00, etc.

Do we think that the schedule should offer this level of consistency, or
is the natural global splay that leaving the value without significant
initialization desired?