explicitly call out OS compatibility at top of OS X install doc
Reordered install types, Pointed to the CI build status as an indication
of platform support, moved up the line about pkg/lib dependencies
to be after the pkg section. Someone stop Atom from tacking newlines to
the ends of files on save, GitHub, you’re not the boss of me!