Prioritize reading node_invalid over error for host re-enrollment
A TLS server might like to return `"node_invalid": true` and an error string
indicating why the node was invalid. The old behavior caused `node_invalid` to
be ignored if `error` was specified, therefore making it impossible for
osqueryd to resolve the error.

With the new behavior, osqueryd can now re-enroll when a reason is provided.