C++ conformance fixes (MSVC /permissive-)
We (the MSVC compiler team) use the OSQuery project as part of our
"Real world code" tests.  I noticed a few places in your windows
specific code where your build would fail if you used the
/permissive- switch because of language conformance issues.
for more information on /permissive- see our blogpost:
https://blogs.msdn.microsoft.com/vcblog/2016/11/16/permissive-switch/
I have not added/modified any tests because this does not change
behavior.

There are two types of language conformance issues fixed in this:

1) Strict string conversion (this is also covered by the /Zc:strictStrings flag)
The 'const' is not implicitly dropped by string literals, which means the following is not allowed:

    //BSTR is a wchar_t*
    BSTR str = L"const string literal"; //error: cannot convert a 'const wchar_t*' to a 'wchar_t*'

I preserved your current behavior by adding a cast to BSTR.  A better fix
would probably be to use SysAllocString (see the documentation for BSTR on MSDN).
I did not make that change because of my limited testing capabilities.

2)Fully qualified inline declarations members inside class

    struct A {
        void A::f() { } // Error: illegal qualified name in member declaration, remove redundant 'A::' to fix
    };