Fix crash caused by boost's unhandled exception
zwass caught that boost has a bug where it throws an exception when we aren't expecting it to. This catches that error and prevent osquery from crashing. It does not address what might be a logic bug in osquery dropping privileges before trying to read the plist. @theopolis can you weigh in on this?