Fix #3220 Error loading packs not verbose enough
If you try to load a pack that doesn't exist, osqueryctl (and config.cpp in general) throws a JSON error because parsing an empty string fails. This adds an extra check for that and throws a more verbose error letting them know that there was a problem getting file contents. Also make the JSON parsing error more verbose so if there is a parsing error they know where to look.