Fix C99 array errors by moving array to the heap
osquery's build (especially tidy) was complainging the dynamically sized stack arrays are a C99 feature. This diff makes the compiler happy by moving them to heap with a malloc.

After they are used, the memory is freed. No exceptions should be raised or caught, and there are no other code paths so this memory should always be freed here. Also tested with ASAN and there were no memory errors announced.