Expand Carver APIs and add Compression
Carve provides a lot of useful functionality to get it's job done like the ability to create archives out of files. This functionality however was contained in the carver class and not easily accessible for general use. This moves that function out of the carver class and restructures it to be more general.

The other thing is that archive compression is currently broken and as far as we can tell will be for the foreseeable future. We need some kind of compression for Carver so I implemented zstd, the super fast, super efficient compression algorithm from Facebook. Now carves are grouped into a tar archive and then compressed with zstd before transmission. zstd benefits are well documents on their github page.