auditd_file_events: Added a new table (see details).
If we want to filter events based on paths, we need the 'realpath'
used to access the file. Problem is that the 'open' syscall may
receive a relative path; we solve this by also listening for CWD
audit events.

The code will now always attempt to canonicalize paths when
possible, and just normalize it when an error occurs. The
issue here is that for the whole procedure to work correctly,
we need to have access to the filesystem in the same exact
state that generated the event (this is the same issue that
the 'realpath' command has).