[RFC] Building on High Sierra
These are all the changes that I had to make to the osquery build system to get make deps and make working on High Sierra. I had to build bison myself to apply a patch addressing a change in macOS' printf stack. Other than that I had to add ninja and reorder a couple dependencies.