AuditEventPublisher: Avoid copying data between threads and more (see details).
The thread responsible for processing the events was initially
performing a copy of the queue filled by the even receiver.

It is now avoiding this by swapping pointers to two different
queue objects.

Some variable names have been changed to be more meaningful, and I
also added some Doxygen comments.