AuditEventPublisher/AuditFim: Various fixes (see details).
The AuditFim class is now discarding events that have been generated
by osquery.

The AuditEventPublisher was still subscribing to the AuditNetlink
component even when the ::setUp() method was returning '1'. This
should be supposed to prevent the ::configure() method from being
called (because you have disabled the publisher via command line)
but it wasn't working.