Nits, safety and type detection
Addressed all of @muffin's nits as well as added logic that will bail on compression if the file grows in size during compression or decompression preventing a possible infinite loop. I also modified the test python server to not print out the request of a carve block to stop it from outputting the whole file to the screen as well as adding type checking to the carved file. If the file is a zst it will now save it with that extention automatically, else it will use tar.