Rearchitecture of OpenBSM eventing
Now we pass a vector of tokens to the subscribers and they can parse them out how they want. Creating a data structure to hold all the information for all event types was not going to scale.

This also adds a new table for SSH logins to the local host, powered by OpenBSM.