Update remote.md
Original: The certificate being passed as parameter int he last example of osqueryd CLI flag __tls_sever_certs was test_server_ca.pem.
Change: I replaced it with test_server.pem.
Reason: As it is mentioned that the flag should pin to the root certificate of server and in the toy example, CN=localhost in subject of certificate. However if we open test_server_ca.pem using openssl, CN=osquery-unittests-ca. Hence the certificate passed is wrong and server certificate should be passed.