oneshot: Support for one-shot queries via packs.
This change tries to addresses couple of things:
* Sometimes a query needs to be run immediately. Without using a small
  interval this is currently not possible.
* Some of the queries don't really need to be re-run.
  For example: Hardware related tables etc.

This functionality seems to be easy to implement using query 'interval'.
When interval is set to 0, it is interpreted as one shot query. It is
scheduled to run once in next 60 seconds. 60 seconds is currently not
configurable but changed. Once the one shot query runs, it is marked as
executed in the database and will not be launched again.