Fix #undef GetObject workaround
On Windows GetObject is defined as a macro and conflicts with the
GetObject method from rapidjson. To fix this we #undef GetObject, but
this needs to be done also right before including rapidjson code. Be
sure to include <osquery/system.h> there which should do the undef.
Remove an unnecessary include as well.