Fix performance regression on process_open_sockets
PR #4024 introduced a performance regression causing a tremendous
increase on the table query time. On a busy server this could by 2 to 3
orders of magnitued higher.

The issue lies with the way socket information is queried. #4024 went
from querying the socket information at once from /proc/net to querying
this information per process from /proc/<pid>/net. This cause a massive
increase on the number of lines that need to be process and therefore
the regression.

This commit builds on top of #4024 but reimplements the logic to query
/proc/net/* files only once per namespace.