Use correct family on rocksdb
Fix RocksDBDatabasePlugin::getHandleForColumnFamily to return the
correct handle for a family. This method was returning handle n-1 for
familiy n due to using kDomains to search for the family index which
misses the default family.

Because this method is always used to get the handle, everything works
but data is written to the DB using the wrong family, e.g.
configurations entries are written to default and queries to
configurations.