[darwin][system_controls] Handle macOS type edge cases
Examining the source to `sysctl` on macOS, I see that there are cases where a sysctl is registered with a type of 0, relying on the format string to express the type, and also cases where it's registered as CTLTYPE_INT but the format string might specify that it's a long int. Handling those.