Pass name of query to lambda in the args of Config::scheduledQueries() by value
Otherwise it is a bit confusing. Const reference to the query is permanent and
user can store it and use outside of lambda, but not the same for a query name.
The reference to the name is a reference to an object on the stack.
And this patch is to explain to user the difference between two arguments:
  - name should be copied by value,
  - query can be used by const ref.