Build posix with -fno-limit-debug-info
When building with lto clang will not generate debug information for
libstdc++ with the assumption we will have those symbols separately.
Since we're statically linking the lib we need to tell clang to generate
these to make debugging easier. One major advantage is to be able to
look into std::string while debugging osquery.