Implement make format_check
The tool tools/formatting/format-check.py is a wrapper around
tools/formatting/git-clang-format.py which (1) caculates the merge-base
between your current branch and master or the passed branch (2) runs
git-clang-format.py with the result and (3) analyses the output and
exits with an appropriate code, printing an message and if necessary the
diff representing the changes required.

This will make code auditing performance code format checks only for the
commits on your feature branch, so they won't be affected in case master
advanced.