Introduce .types shell command
The OSQuery API provides a method (`getQueryColumns`) for getting
the column types of a query. It's far from perfect (there are
lots of cases where it can't tell, or even gets it wrong), so
it's sometimes useful to be able to see what it thinks about
a particular query.

This commit introduces a shell command `.types` that takes a query and
shows what the API would give for its column types.

```
osquery> .types SELECT name, celsius, celsius - 100 AS degrees_over_boiling FROM temperature_sensors;
+----------------------+--------+
| name                 | type   |
+----------------------+--------+
| name                 | TEXT   |
| celsius              | DOUBLE |
| degrees_over_boiling | BIGINT |
+----------------------+--------+
```