Support CDHash algorithms other than SHA-1
CDHash is not guaranteed to be SHA-1.  In fact, the algorithm is being
changed to SHA-256.  Remove the dependency on SHA-1 from the code in
order to support CDHashes produced with other hash algorithms, but still
rely on the OS to decide which CDHash to select if there are multiple
hashes produced using different hash algorithms.